// src/api/__generated__/story/model/dramaEditEpisodeItem.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Service API
 * Mini Drama Platform
 * OpenAPI spec version: v1.0
 */

export interface DramaEditEpisodeItem {
  id?: number;
  episodeNo?: number;
  title?: string;
  description?: string;
  videoObjectKey?: string;
  durationSec?: number;
  width?: number;
  height?: number;
}