// src/api/__generated__/story/model/updateDramaRequest.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Mini Drama Service API
 * Mini Drama Platform
 * OpenAPI spec version: v1.0
 */
import type { ActorCollectionChanges } from './actorCollectionChanges';
import type { DramaEditEpisodeItem } from './dramaEditEpisodeItem';

export interface UpdateDramaRequest {
  uploadSessionId?: number;
  /**
   * @minLength 0
   * @maxLength 200
   */
  title?: string;
  /**
   * @minLength 0
   * @maxLength 1000
   */
  description?: string;
  coverObjectKey?: string;
  episodes?: DramaEditEpisodeItem[];
  actorCollectionChanges?: ActorCollectionChanges;
  tagIds?: number[];
}